Streaming Bot/requirements.txt
langchain===0.0.184
streamlit
langchain-openai
python-dotenv